+++ src/Input/Input.tsx
@@ -1,8 +1,10 @@
-import { forwardRef, InputHTMLAttributes } from 'react'
+import { forwardRef, InputHTMLAttributes, useState } from 'react'
 
 import clsx from 'clsx'
 
 import styles from './Input.module.scss'
+import EyeClosed from '../../icons/system/eye-closed.svg'
+import Eye from '../../icons/system/eye.svg'
 
 type InputProps = Omit<InputHTMLAttributes<HTMLInputElement>, 'children'> & {
   label?: string
@@ -16,6 +18,7 @@ type InputProps = Omit<InputHTMLAttributes<HTMLInputElement>, 'children'> & {
   inputClassName?: string
   hintClassName?: string
   errorClassName?: string
+  required?: boolean
 }
 
 const Input = forwardRef<HTMLInputElement, InputProps>(
@@ -33,27 +36,54 @@ const Input = forwardRef<HTMLInputElement, InputProps>(
       inputClassName,
       hintClassName,
       errorClassName,
+      type,
+      required = false,
       ...props
     },
     ref
   ) => {
+    // State to manage password visibility
+    const [showPassword, setShowPassword] = useState(false)
+
+    // Toggle password visibility
+    const togglePasswordVisibility = () => {
+      setShowPassword((prev) => !prev)
+    }
+
+    // Determine the input type
+    const inputType = type === 'password' && showPassword ? 'text' : type
     return (
       <div title={tooltip} className={clsx(styles.inputWrapper, className)}>
         {label && (
           <label htmlFor={id} className={clsx(styles.label, labelClassName)}>
             {label}
+            {required && <span className={styles.required}>*</span>}
           </label>
         )}
-        <input
-          id={id}
-          ref={ref}
-          className={clsx(styles.input, inputClassName, {
-            [styles.inputError]: isError,
-            [styles.inputSuccess]: isSuccess && !isError,
-          })}
-          aria-invalid={isError}
-          {...props}
-        />
+        <div className={styles.inputContainer}>
+          <input
+            id={id}
+            ref={ref}
+            type={inputType}
+            className={clsx(styles.input, inputClassName, {
+              [styles.inputError]: isError,
+              [styles.inputSuccess]: isSuccess && !isError,
+            })}
+            aria-invalid={isError}
+            {...props}
+          />
+          {/* Show/hide password icon for password inputs */}
+          {type === 'password' && (
+            <button
+              type="button"
+              onClick={togglePasswordVisibility}
+              className={styles.toggleButton}
+              aria-label={showPassword ? 'Hide password' : 'Show password'}
+            >
+              {showPassword ? <EyeClosed className={styles.eyeIcon} /> : <Eye className={styles.eyeIcon} />}
+            </button>
+          )}
+        </div>
         {error ? (
           <span className={clsx(styles.error, errorClassName)}>{error}</span>
         ) : (