+++ src/OTPInput/index.ts
@@ -1,1 +1,3 @@
 export { default } from './OTPInput'
+
+export type { OTPInputProps } from './OTPInput'